Add: new branch develop, and delete UserRoles and AppointmentServices

+++ migrations/20230606233048-create-service.js
@@ -14,7 +14,7 @@ module.exports = {
         unique: true
       },
       price: {
-        type: Sequelize.FLOAT
+        type: Sequelize.INTEGER,
       },
       duration: {
         type: Sequelize.STRING
